+++ src/components/AdminLoginPage/index.js
@@ -8,7 +8,7 @@ import LandingFooter from '../LandingFooter';
 import InputText from '../InputText';
 import InputPassword from '../InputPassword';
 import PrimaryButton from '../PrimaryButton';
-import Spinner from '../SpinnerComponent';
+import Spinner from '../Spinner';
 import { API_BASE_URL } from '../../config';
 import '../LoginPage/LoginPage.css';
