+++ src/components/NodesList/index.jsx
@@ -6,7 +6,7 @@ import tellAge from '../../helpers/ageUtility';
 import './NodesList.css';
 import Header from '../Header';
 import Status from '../Status';
-import { BigSpinner } from '../SpinnerComponent';
+import Spinner from '../Spinner';
 import InformationBar from '../InformationBar';
 import SideNav from '../SideNav';
 
@@ -77,7 +77,7 @@ class NodesList extends Component {
                         <tr className="TableLoading">
                           <td>
                             <div className="SpinnerWrapper">
-                              <BigSpinner />
+                              <Spinner />
                             </div>
                           </td>
                         </tr>
