+++ src/components/Spinner/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import './Spinner.css';
+
+
+const Spinner = (props) => {
+  const { size } = props;
+
+  return (
+    <div className="SmallSpinnerWrapper">
+      <div className={`Spinner Spinner${size}`} />
+    </div>
+  );
+};
+
+export default Spinner;
