--- src/components/SpinnerComponent/SpinnerComponent.css
@@ -1,72 +1,0 @@
-/*! SMALL SPINNER */
-.SmallSpinnerWrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.Spinner {
-    border: 2px solid white;
-    width: 1rem;
-    height: 1rem;
-    border-radius: 100%;
-    border-top: 2px solid #f7b21f;
-    animation: spin 600ms linear infinite;
-}
-
-.SpinnerBig {
-  width: 4rem;
-  height: 4rem;
-}
-
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-    100% {
-        transform: rotate(360deg);
-    }
-}
-
-/*! BIG SPINNER */
-/* Here is where the magic happens */
-.BigSpinnerContainer {
-  width: 100%;
-  height: 100%;
-  --offset: 187;
-  --duration: 1.4s;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-
-.BigSpinner {
-  animation: rotator var(--duration) linear infinite;
-}
-
-@keyframes rotator {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
-}
-
-.SpinnerPath {
-  stroke-dasharray: var(--offset);
-  stroke-dashoffset: 0;
-  transform-origin: center;
-  /* stroke: #008AC1; */
-  stroke: #f7b21f;
-  animation: dash var(--duration) ease-in-out infinite;
-}
-
-@keyframes dash {
- 0% { stroke-dashoffset: var(--offset); }
- 50% {
-   stroke-dashoffset: var(--offset)/4;
-   transform:rotate(135deg);
- }
- 100% {
-   stroke-dashoffset: var(--offset);
-   transform:rotate(450deg);
- }
-}
